+++ mysql-operator/.ci/build.sh
@@ -3,7 +3,8 @@
 set -e
 
 cd $GALLY_PROJECT_ROOT
-export AGENT_VERSION=$(gally list -p agent | grep GALLY_PROJECT_VERSION | cut -d'"' -f4)
-docker build --build-arg agent_version=$AGENT_VERSION -t $TAG:$GALLY_PROJECT_VERSION .
+make test
+export AGENT_CODE=$(grep "DefaultAgentVersion =" main.go | cut -d '"' -f2)
+docker build --build-arg agent_version=$AGENT_CODE -t $TAG:$GALLY_PROJECT_VERSION .
 docker push $TAG:$GALLY_PROJECT_VERSION
 cd $GALLY_ROOT